requirements.txt
openai
python-dotenv
bcrypt